requirements.txt
pip~=21.2.4
toml~=0.10.2
zlib~=1.2.12
wheel~=0.37.1
brotli~=1.0.9
openssl~=1.1.1w
cryptography~=3.4.8
py~=1.11.0
lxml~=4.8.0
tornado~=6.1
keyring~=23.4.0
django~=4.2
asgiref~=3.6.0
sqlparse~=0.4.3
psycopg2~=2.9.3
colorama~=0.4.4
yaml~=0.2.5
pyyaml~=6.0
jinja2~=2.11.3
tblib~=1.7.0
numpy~=1.21.5
requests~=2.27.1
dataclasses~=0.8
docutils~=0.17.1
typing_extensions~=4.1.1
djangorestframework~=3.14.0
setuptools~=61.2.0
cython~=0.29.28
future~=0.18.2
inflection~=0.5.1
markdown~=3.3.4
pygments~=2.11.2
pytz~=2021.3
pyjwt~=2.1.0
rsa~=4.7.2
pyasn1~=0.4.8
aiohttp~=3.8.1
cachetools~=4.2.2
certifi~=2024.8.30
urllib3~=1.26.9
packaging~=21.3
cffi~=1.15.0
pyparsing~=3.0.4
idna~=3.3
chardet~=4.0.0
simplejson~=3.17.6
werkzeug~=2.0.3
click~=8.1.7